trading_api/requirements.txt
fastapi
uvicorn
yfinance
pandas
polars
plotly
# MetaTrader5  # Windows only - not needed for Yahoo Finance on macOS
psycopg2-binary
python-dotenv
requests
